chore: 기존 쿠폰 모달 삭제
